# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Bolo-Man API
  description: Cameroon-focused Daily Services Platform API
  version: '1.0'
  contact:
    name: Bolo-Man Support
    email: support@boloman.cm
servers:
  - url: https://api.boloman.cm
    description: Production
  - url: http://localhost:3000
    description: Development
security:
  - bearerAuth: []
paths:
...
  /providers:
    get:
      tags: [providers]
      summary: Search/list providers
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: minRating
          in: query
          schema: { type: number }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        200:
          description: List of providers
  /providers/profile:
    get:
      tags: [providers]
      summary: Get provider profile (for logged-in provider)
      responses:
        200:
          description: Provider profile

    put:
      tags: [providers]
      summary: Update provider profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                businessName: { type: string }
                description: { type: string }
                categories: { type: array, items: { type: string } }
                availability: { type: object }
                maxDistanceKm: { type: number }
...
  /services:
    get:
      tags: [services]
      summary: List services
      parameters:
        - name: categoryId
          in: query
          schema: { type: string }
      responses:
        200:
          description: List of services
...
  /payments:
    post:
      tags: [payments]
      summary: Initiate a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId: { type: string }
                amount: { type: integer, minimum: 100 }
                currency: { type: string, default: 'XAF' }
                method: { type: string, enum: [MOMO, ORANGE, CARD, COD, WALLET] }

    get:
      tags: [payments]
      summary: Get payment history
      responses:
        200:
          description: Payment history
...
  /payments/webhooks/momo:
    post:
      tags: [payments]
      summary: MoMo webhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
...
  /reviews:
    post:
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
  /reviews/provider/{providerId}:
    get:
      tags: [reviews]
      summary: Get provider reviews
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Reviews list
  /reviews/{id}/response:
    put:
      tags: [reviews]
      summary: Respond to review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                response: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...